% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aozora.R
\name{aozora}
\alias{aozora}
\title{Download text file from Aozora Bunko}
\usage{
aozora(url = NULL, txtname = NULL, directory = "cache")
}
\arguments{
\item{url}{URL of text download link.}

\item{txtname}{new file name as which text is saved.}

\item{directory}{directory name where new file is saved.}
}
\value{
the path to the new file
}
\description{
download a file from specified Aozora Bunko URL and unzip the file,
convert it to UTF-8 format.
}
\seealso{
\url{https://gist.github.com/ishida-m/7969049}
}
