% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AozoraBunkoSnapshot}
\alias{AozoraBunkoSnapshot}
\title{Meta data of text files published on Aozora Bunko}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16193 rows and 55 columns.
}
\source{
\url{http://www.aozora.gr.jp/index_pages/list_person_all_extended_utf8.zip}
}
\usage{
AozoraBunkoSnapshot
}
\description{
Meta data of text files published on Aozora Bunko
}
\seealso{
\url{https://qiita.com/ksato9700/items/48fd0eba67316d58b9d6}
}
\keyword{datasets}
