% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmecab.R
\name{cmecab_c}
\alias{cmecab_c}
\title{Call CMECAB Tagger}
\usage{
cmecab_c(str = "", opt = "", sep = " ")
}
\arguments{
\item{str}{String to be tokenized}

\item{opt}{String to be passed as tagger options (ex. "--partial").}

\item{sep}{String used as separator.}
}
\value{
a list
}
\description{
Call CMECAB Tagger
}
