% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-for-csv.R
\name{csvutil_join}
\alias{csvutil_join}
\title{Utility for handling CSV-like string}
\usage{
csvutil_join(chr)
}
\arguments{
\item{chr}{Character vector.}
}
\value{
Character scalar.
}
\description{
Equivalent to \code{paste(chr, collapse = ",")}.
}
