% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-for-csv.R
\name{csvutil_tokenize}
\alias{csvutil_tokenize}
\title{Utility for handling CSV-like string}
\usage{
csvutil_tokenize(str, max = 99L)
}
\arguments{
\item{str}{Character scalar.}

\item{max}{Integer (column size limit).}
}
\value{
Character vector.
}
\description{
Parse a CSV-row-like string something like \code{"a,b,c"}.
}
