% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{emojiRegexp}
\alias{emojiRegexp}
\title{Return emoji subset}
\usage{
emojiRegexp(version = c(6, 7, 8))
}
\arguments{
\item{version}{version digit.}
}
\value{
character representing regular expression that mactches emoji in Unicode code point order.
}
\description{
Return emoji subset
}
