% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvutil.R
\name{escape}
\alias{escape}
\title{Util for Handling CSV-like string}
\usage{
escape(str)
}
\arguments{
\item{str}{character scalar.}
}
\value{
character scalar.
}
\description{
escape space, tab, `"` and `,` in string
}
