% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExtendedLettersJp}
\alias{ExtendedLettersJp}
\title{List of letters known as "Kishu-ison-moji"}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 457 rows and 2 columns.
}
\usage{
ExtendedLettersJp
}
\description{
List of letters known as "Kishu-ison-moji"
}
\keyword{datasets}
