% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fastestword}
\alias{fastestword}
\title{Call mecab command directly}
\usage{
fastestword(
  chr = c(""),
  outfile = file.path(getwd(), "output.txt"),
  encoding = "UTF-8",
  opt = "-Owakati",
  ...
)
}
\arguments{
\item{chr}{character vector.}

\item{outfile}{path to a file that MeCab writes output.}

\item{encoding}{encoding of tempfile that MeCab reads.}

\item{opt}{options passed to mecab command.}

\item{...}{any other arguments are passed to \code{base::writeLines()}.}
}
\value{
the function returns the value passed to `outfile` argument invisibly
}
\description{
call mecab command directly via \code{base::system()}.
}
\details{
It is useful especially when just tokenizing loads of text.
Since `mecab -Owakati` command is specially-tuned,
it is generally faster that you directly call mecab command
than programmatically use mecab tagger
if you would like to just tokenize (wakachi-gaki) texts.
}
