% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format}
\alias{format}
\title{Format cmecab_c() output}
\usage{
format(list, sep = " ")
}
\arguments{
\item{list}{List that is output from cmecab_c()}

\item{sep}{String used as separator with which it replaces tab.}
}
\value{
a tibble
}
\description{
Format cmecab_c() output
}
