% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_blank}
\alias{is_blank}
\title{Check if scalars are blank}
\usage{
is_blank(x, trim = TRUE, ...)
}
\arguments{
\item{x}{Object to check its emptiness.}

\item{trim}{Boolean.}

\item{...}{Additional arguments for \code{base::sapply()}.}
}
\value{
Logical values.
}
\description{
Check if scalars are blank
}
\details{
This function is a port of \code{rapportools::is.empty()}.
Thus the source of this function is licensed under AGPL-3.
}
\seealso{
\url{https://github.com/Rapporter/rapportools/blob/master/R/utils.R#L334}
}
