% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_dyn_available.R
\name{is_dyn_available}
\alias{is_dyn_available}
\title{Check if the mecab dynamic library is available}
\usage{
is_dyn_available(dynlib = "libmecab")
}
\arguments{
\item{dynlib}{Default value is 'libmecab'.}
}
\value{
Logical.
}
\description{
Check if the mecab dynamic library is available
}
