% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mecab_available.R
\name{is_mecab_available}
\alias{is_mecab_available}
\title{Check if the mecab or its dynamic library is available}
\usage{
is_mecab_available()
}
\value{
Logical.
}
\description{
Check if the mecab or its dynamic library is available
}
