% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvutil.R
\name{join}
\alias{join}
\title{Util for Handling CSV-like string}
\usage{
join(char)
}
\arguments{
\item{char}{character vector.}
}
\value{
character scalar
}
\description{
equivalent to \code{paste(char, collapse = ",")}
}
