% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MiyazawaKenji}
\alias{MiyazawaKenji}
\title{List of Shinji-Shinkana works written by Miyazawa Kenji on Aozora Bunko}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 83 rows and 2 columns.
}
\usage{
MiyazawaKenji
}
\description{
List of Shinji-Shinkana works written by Miyazawa Kenji on Aozora Bunko
}
\keyword{datasets}
