% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize}
\alias{normalize}
\title{Normalize CJK text}
\usage{
normalize(str)
}
\arguments{
\item{str}{Character vector.}
}
\value{
Normalized text.
}
\description{
Normalize text into neologd-style.
}
\seealso{
\url{https://github.com/neologd/mecab-ipadic-neologd/wiki/Regexp}
}
