% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OneLettersJp}
\alias{OneLettersJp}
\title{List of Japanese letters (Hiragana and Katakana)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 169 rows and 2 columns.}
\usage{
OneLettersJp
}
\description{
List of Japanese letters (Hiragana and Katakana)
}
\keyword{datasets}
