% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify.R
\name{prettify}
\alias{prettify}
\title{Prettify tokenized output}
\usage{
prettify(list, sep = " ", into = get_dict_features("ipa"))
}
\arguments{
\item{list}{List that comes out of \code{rjavacmecab::cmecab()}
or \code{rjavacmecab::igo(mode = "parse")}.}

\item{sep}{Character scalar that is used as separators
with which the function replaces tab.}

\item{into}{Character vector that is used as column names of
features.}
}
\value{
data.frame.
}
\description{
Prettify tokenized output
}
