% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igo.R
\name{rebuild_igo_tagger}
\alias{rebuild_igo_tagger}
\title{Rebuild Igo tagger}
\usage{
rebuild_igo_tagger(data_dir = .pkgenv[["igodic"]])
}
\arguments{
\item{data_dir}{Character vector.}
}
\value{
The stored tagger instance is returned invisibly.
}
\description{
Rebuild an instance of igo tagger using provided options.
The igo tagger instance is stored in the package internal environment.
}
\keyword{internal}
