% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjavacmecab.R
\docType{package}
\name{rjavacmecab}
\alias{rjavacmecab}
\alias{rjavacmecab-package}
\title{rjavacmecab: rJava Interface to CMeCab}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

A simple rJava interface to cmecab-java.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/rjavacmecab}
  \item Report bugs at \url{https://github.com/paithiov909/rjavacmecab/issues}
}

}
\author{
\strong{Maintainer}: Kato Akiru \email{paithiov909@gmail.com}

}
