% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StopWordsJp}
\alias{StopWordsJp}
\title{Japanese stopwords list (Slothlib)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 310 rows and 2 columns.
}
\source{
\url{http://svn.sourceforge.jp/svnroot/slothlib/CSharp/Version1/SlothLib/NLP/Filter/StopWord/word/Japanese.txt}
}
\usage{
StopWordsJp
}
\description{
Japanese stopwords list (Slothlib)
}
\keyword{datasets}
