% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvutil.R
\name{tokenize}
\alias{tokenize}
\title{Util for Handling CSV-like string}
\usage{
tokenize(str, max = 99L)
}
\arguments{
\item{str}{character scalar..}

\item{max}{integer. the limitation of column size.}
}
\value{
character vector
}
\description{
parse a CSV-row-like string something like `"a,b,c"`
}
