% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aozora.R
\name{aozora}
\alias{aozora}
\title{Download text file from Aozora Bunko}
\usage{
aozora(url, txtname = NULL, directory = file.path(getwd(), "cache"))
}
\arguments{
\item{url}{URL of text download link.}

\item{txtname}{New file name as which text is saved.
If \code{NULL} provided, keeps name of the source file.}

\item{directory}{Path where new file is saved.}
}
\value{
The path to the new file.
}
\description{
Download a file from specified URL, unzip the file
and convert it to UTF-8.
}
\seealso{
\url{https://gist.github.com/ishida-m/7969049}
}
