% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aozora.R
\name{aozora}
\alias{aozora}
\title{Download text file from Aozora Bunko}
\usage{
aozora(url = NULL, txtname = NULL, directory = "cache")
}
\arguments{
\item{url}{URL of text download link.}

\item{txtname}{New file name as which text will be saved.}

\item{directory}{Direcotry name where new file will be saved.}
}
\value{
the path to the new file
}
\description{
Download a file from specified Aozora Bunko URL and unzip the file,
convert it to UTF-8 format.
}
\seealso{
\url{https://gist.github.com/ishida-m/7969049}
}
