% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmecab.R
\name{cmecab}
\alias{cmecab}
\title{Call CMeCab tagger}
\usage{
cmecab(str = "", opt = "", sep = " ")
}
\arguments{
\item{str}{character vector to be tokenized.}

\item{opt}{character scalar to be passed as tagger options (ex. "-d").}

\item{sep}{character scalar to be used as separator
with which the function replaces tab.}
}
\value{
a list
}
\description{
Call CMeCab tagger
}
