% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmecab.R
\name{cmecab}
\alias{cmecab}
\title{Call CMeCab tagger}
\usage{
cmecab(chr, opt = "", split = FALSE, mode = c("parse", "wakati"))
}
\arguments{
\item{chr}{Character vector to be tokenized.}

\item{opt}{Character scalar to be passed as tagger options (ex. "-d").}

\item{split}{Logical. If true (by default), the function splits character vector
into sentences using \code{stringi::stri_split_boundaries(type = "sentence")} before analyzing them.}

\item{mode}{Character scalar.}
}
\value{
A tibble.
}
\description{
Call CMeCab tagger
}
