% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmecab.R
\name{cmecab}
\alias{cmecab}
\title{Calls CMeCab tagger}
\usage{
cmecab(chr, opt = "", sep = " ")
}
\arguments{
\item{chr}{Character vector to be tokenized.}

\item{opt}{Character scalar to be passed as tagger options (ex. "-d").}

\item{sep}{Character scalar to be used as separator
with which the function replaces tab.}
}
\value{
List.
}
\description{
Calls CMeCab tagger
}
