% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmecab.R
\name{cmecab_c}
\alias{cmecab_c}
\title{Call CMeCab tagger}
\usage{
cmecab_c(str = "", opt = "", sep = " ")
}
\arguments{
\item{str}{String to be tokenized}

\item{opt}{String to be passed as tagger options (ex. "-d").}

\item{sep}{String used as separator with which it replaces tab.}
}
\value{
a list
}
\description{
Call CMeCab tagger
}
