% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-for-csv.R
\name{csvutil_escape}
\alias{csvutil_escape}
\title{Utility for handling CSV-like string}
\usage{
csvutil_escape(chr)
}
\arguments{
\item{chr}{Character vector.}
}
\value{
Character vector.
}
\description{
Escape space, tab, \verb{"} and \verb{,} in strings.
}
