% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-for-csv.R
\name{escape}
\alias{escape}
\title{Utility for handling CSV-like string}
\usage{
escape(str)
}
\arguments{
\item{str}{Character scalar.}
}
\value{
Character scalar.
}
\description{
Escape space, tab, \verb{"} and \verb{,} in string.
}
