% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvutil.R
\name{escape}
\alias{escape}
\title{Utility for handling CSV-like string}
\usage{
escape(str)
}
\arguments{
\item{str}{Character scalar.}
}
\value{
Character scalar.
}
\description{
Escapes space, tab, `"` and `,` in string.
}
