% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastestword.R
\name{fastestword}
\alias{fastestword}
\title{Call mecab command directly}
\usage{
fastestword(chr = c(""), outfile = file.path(getwd(), "output.txt"),
  encoding = "UTF-8", opt = "-Owakati", ...)
}
\arguments{
\item{chr}{character vector.}

\item{outfile}{fullpath to a file that MeCab will write output.}

\item{encoding}{encoding of tempfile that MeCab will read.}

\item{opt}{options passed to mecab command.}

\item{...}{arguments passed to writeLines().}
}
\value{
return the value pased to `outfile` argument invisibly
}
\description{
Call mecab command directly via system().
}
\details{
It is useful especially when just tokenizing loads of text.
Since `mecab -Owakati` command is specially-tuned,
it is generally faster that you directly call mecab command
than programmatically use mecab tagger
if you would like to just tokenize (wakachi-gaki) texts.
}
