% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igo.R
\name{igo}
\alias{igo}
\title{Call Igo tagger}
\usage{
igo(chr, split = TRUE, mode = c("parse", "wakati"))
}
\arguments{
\item{chr}{Character vector to be tokenized.}

\item{split}{Logical. If true (by default), the function splits character vector
into sentences using \code{stringi::stri_split_boudaries(type = "sentence")}
before analyzing them.}

\item{mode}{Character scalar.}
}
\value{
A tibble.
}
\description{
Call Igo tagger
}
\details{
This function is a wrapper of Igo (v0.4.5) that is a Java reimplementaion of MeCab.
The Igo tagger has 2 methods, which are 'parse' and 'wakati', and both is available from this function.
}
