% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvutil.R
\name{join}
\alias{join}
\title{Utility for handling CSV-like string}
\usage{
join(chr)
}
\arguments{
\item{chr}{Character vector.}
}
\value{
Character scalar.
}
\description{
Equivalent to \code{paste(chr, collapse = ",")}.
}
