% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ngram_tokenizer}
\alias{ngram_tokenizer}
\title{Ngrams tokenizer}
\usage{
ngram_tokenizer(n = 1L, skip_word_none = TRUE, locale = NULL)
}
\arguments{
\item{n}{Integer.}

\item{skip_word_none}{Boolean.}

\item{locale}{Character scalar, NULL or "" for default locale.}
}
\value{
N-gram tokenizer function.
}
\description{
Makes N-gram tokenizer function.
}
\seealso{
\url{https://rpubs.com/brianzive/textmining}
}
