% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prettify}
\alias{prettify}
\title{Prettify cmecab output}
\usage{
prettify(list, sep = " ")
}
\arguments{
\item{list}{List that comes out of \code{rjavacmecab::cmecab()}.}

\item{sep}{Character scalar that is used as separator
with which the function replaces tab.}
}
\value{
data.frame.
}
\description{
Prettify cmecab output
}
