% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prettify}
\alias{prettify}
\title{Prettify cmecab() output}
\usage{
prettify(list, sep = " ")
}
\arguments{
\item{list}{List that is output from cmecab().}

\item{sep}{String used as separator with which it replaces tab.}
}
\value{
dataframe
}
\description{
Prettify cmecab() output
}
