% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmecab.R
\name{rebuild_tagger}
\alias{rebuild_tagger}
\title{Rebuild CMeCab tagger}
\usage{
rebuild_tagger(opt = "")
}
\arguments{
\item{opt}{Character vector.}
}
\value{
The stored tagger instance is returned invisibly.
}
\description{
Rebuild an instance of standard tagger using provided options.
The standard tagger instance is stored in the package internal environment.
}
