% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmecab.R
\name{rebuild_tagger}
\alias{rebuild_tagger}
\title{Rebuild CMeCab tagger}
\usage{
rebuild_tagger(opt = "")
}
\arguments{
\item{opt}{character vector.}
}
\value{
return the stored tagger instance invisibly.
}
\description{
Rebuild instance of tagger using provided options.
}
