% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjavacmecab-package.R
\docType{package}
\name{rjavacmecab}
\alias{rjavacmecab}
\alias{rjavacmecab-package}
\title{rjavacmecab: rJava Interface to CMeCab}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A simple rJava interface to cmecab-java.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paithiov909.github.io/rjavacmecab/}
  \item \url{https://github.com/paithiov909/rjavacmecab}
  \item Report bugs at \url{https://github.com/paithiov909/rjavacmecab/issues}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
\keyword{internal}
