% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-for-csv.R
\name{tokenize}
\alias{tokenize}
\title{Utility for handling CSV-like string}
\usage{
tokenize(str, max = 99L)
}
\arguments{
\item{str}{Character scalar.}

\item{max}{Limitation of column size.}
}
\value{
Character vector
}
\description{
Parses a CSV-row-like string something like \code{"a,b,c"}.
}
