% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvutil.R
\name{tokenize}
\alias{tokenize}
\title{Util for Handling CSV-like string}
\usage{
tokenize(str, max = 99L)
}
\arguments{
\item{str}{Character scalar.}

\item{max}{Limitation of column size.}
}
\value{
Character vector
}
\description{
Parse a CSV-row-like string something like `"a,b,c"`.
}
